Convert diversions to list instead of sparse array, part 1. * m4/m4module.h (m4_shipout_text, m4_shipout_string): Use size_t for length. * m4/output.c (diversion, total_buffer_size, output_unused): Track size in size_t.  All users changed. (struct diversion): Reduce size now, to compensate for part 2. (m4_make_diversion, make_room_for): Avoid malloc overflow. (m4_output_exit): Fix typo in assert. (m4_output_init, make_room_for, m4_make_diversion) (m4_insert_diversion, m4_freeze_diversions): Adjust users of struct m4_diversion to disambiguate the new union. * tests/builtins.at (divert): Detect this failure.

+++ m4/m4module.h
@@ -413,10 +413,10 @@ extern	void	m4_input_print	(m4 *, m4_obstack *, m4_input_block *);
 
 extern void	m4_output_init	  (m4 *);
 extern void	m4_output_exit	  (void);
-extern void	m4_shipout_text	  (m4 *, m4_obstack *, const char *, int);
+extern void	m4_shipout_text	  (m4 *, m4_obstack *, const char *, size_t);
 extern void	m4_shipout_int    (m4_obstack *, int);
 extern void	m4_shipout_string (m4 *, m4_obstack *, const char *,
-				   int, bool);
+				   size_t, bool);
 
 extern void	m4_make_diversion    (m4 *, int);
 extern void	m4_insert_diversion  (m4 *, int);
